refactor: refactor direct request
