+++ src/components/Payment/Views/Status.payment.view.tsx
@@ -10,19 +10,22 @@ import { paymentActions } from '@/redux/slices/payment-slice'
 import { ApiUser } from '@/services/users'
 import { getInitialsFromName, printableAddress } from '@/utils'
 import { useRouter } from 'next/navigation'
-import { useMemo } from 'react'
+import { useMemo, useState, useEffect } from 'react'
 import { useDispatch } from 'react-redux'
+import Loading from '@/components/Global/Loading'
 
-interface DirectSendSuccessViewProps {
+type DirectSuccessViewProps = {
     user?: ApiUser
     amount?: string
     message?: string
     recipientType?: RecipientType
+    type: 'SEND' | 'REQUEST'
 }
 
-const PaymentStatusView = ({ user, amount, message, recipientType }: DirectSendSuccessViewProps = {}) => {
+const DirectSuccessView = ({ user, amount, message, recipientType, type }: DirectSuccessViewProps) => {
     const router = useRouter()
     const { chargeDetails, parsedPaymentData } = usePaymentStore()
+    const [showCheck, setShowCheck] = useState(false)
     const dispatch = useDispatch()
 
     const recipientName = useMemo(() => {
@@ -40,6 +43,23 @@ const PaymentStatusView = ({ user, amount, message, recipientType }: DirectSendS
     }, [amount, chargeDetails])
 
     const initials = getInitialsFromName(recipientName)
+
+    useEffect(() => {
+        // show loading for a brief moment, then show check mark
+        const checkTimeout = setTimeout(() => {
+            setShowCheck(true)
+        }, 800)
+
+        // redirect to home after 2 seconds
+        const redirectTimeout = setTimeout(() => {
+            router.push('/home')
+        }, 2000)
+
+        return () => {
+            clearTimeout(checkTimeout)
+            clearTimeout(redirectTimeout)
+        }
+    }, [router])
 
     const handleDone = () => {
         // reset payment state when done
@@ -65,7 +85,7 @@ const PaymentStatusView = ({ user, amount, message, recipientType }: DirectSendS
 
                     <div className="space-y-1">
                         <h1 className="text-sm font-bold">
-                            You just sent{' '}
+                            You just {type === 'SEND' ? 'sent' : 'requested'}{' '}
                             {recipientType !== 'USERNAME' ? (
                                 <AddressLink
                                     className="text-sm font-bold text-black no-underline"
@@ -85,11 +105,11 @@ const PaymentStatusView = ({ user, amount, message, recipientType }: DirectSendS
 
             <Button onClick={handleDone} shadowSize="4" className="mx-auto w-38 rounded-full">
                 <div className="flex size-7 items-center justify-center gap-0">
-                    <Icon name="check" size={24} />
+                    {showCheck ? <Icon name="check" size={24} /> : <Loading />}
                 </div>
                 <div>Done!</div>
             </Button>
         </div>
     )
 }
-export default PaymentStatusView
+export default DirectSuccessView
